# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI20K Agent System",
    "description": "AI20K Agent System API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/data-model": {
      "get": {
        "tags": [
          "Data Models"
        ],
        "summary": "Get Current Data Model",
        "description": "Lấy DBML và revision hiện tại của dự án.",
        "operationId": "getDataModel",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa Data Model",
              "title": "Project Id"
            },
            "description": "ID dự án chứa Data Model"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DataModelResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Data Models"
        ],
        "summary": "Update Current Data Model",
        "description": "Lưu trực tiếp DBML do người dùng chỉnh sửa thủ công.",
        "operationId": "updateDataModel",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa Data Model",
              "title": "Project Id"
            },
            "description": "ID dự án chứa Data Model"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDataModelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DataModelResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/data-model/validation-issues": {
      "get": {
        "tags": [
          "Data Models"
        ],
        "summary": "Get Data Model Validation Issues",
        "description": "Trả các lỗi và cảnh báo validation của snapshot DBML hiện tại.",
        "operationId": "getDataModelValidationIssues",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa Data Model",
              "title": "Project Id"
            },
            "description": "ID dự án chứa Data Model"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_DataModelValidationIssueResponse__"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/data-model/validate": {
      "post": {
        "tags": [
          "Data Models"
        ],
        "summary": "Validate Data Model Draft",
        "description": "Kiểm tra draft hiện tại mà không gọi LLM hoặc ghi snapshot.",
        "operationId": "validateDataModelDraft",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa Data Model",
              "title": "Project Id"
            },
            "description": "ID dự án chứa Data Model"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateDataModelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_DataModelValidationIssueResponse__"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/data-model/ddl": {
      "get": {
        "tags": [
          "Data Models"
        ],
        "summary": "Generate Data Model Ddl",
        "description": "Sinh DDL từ revision Data Model hiện hành.",
        "operationId": "generateDataModelDdl",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa Data Model",
              "title": "Project Id"
            },
            "description": "ID dự án chứa Data Model"
          },
          {
            "name": "db_type",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SandboxDbType",
              "default": "POSTGRESQL"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DataModelDdlResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Code",
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông báo lỗi"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Mã lỗi ổn định cho client"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Danh sách chi tiết lỗi theo field"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "error_code"
        ],
        "title": "ApiErrorResponse",
        "description": "Error envelope được Global Exception Handler trả về."
      },
...
      "ApiResponse_DataModelDdlResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataModelDdlResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[DataModelDdlResponse]"
      },
      "ApiResponse_DataModelResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataModelResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[DataModelResponse]"
      },
...
      "ApiResponse_list_DataModelValidationIssueResponse__": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DataModelValidationIssueResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[list[DataModelValidationIssueResponse]]"
      },
...
      "DataModelDdlResponse": {
        "properties": {
          "ddl": {
            "type": "string",
            "title": "Ddl",
            "description": "Script DDL đã sinh"
          },
          "db_type": {
            "$ref": "#/components/schemas/SandboxDbType",
            "description": "Database type đích"
          },
          "data_model_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Data Model Revision",
            "description": "Revision Data Model nguồn"
          }
        },
        "type": "object",
        "required": [
          "ddl",
          "db_type",
          "data_model_revision"
        ],
        "title": "DataModelDdlResponse",
        "description": "DDL sinh từ Data Model hiện hành."
      },
      "DataModelResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "ID của Data Model"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id",
            "description": "ID dự án sở hữu Data Model"
          },
          "dbml": {
            "type": "string",
            "minLength": 1,
            "title": "Dbml",
            "description": "Snapshot DBML hiện tại"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision",
            "description": "Revision phục vụ optimistic locking"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Thời điểm tạo theo ISO 8601"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Thời điểm cập nhật theo ISO 8601"
          },
          "is_outdated": {
            "type": "boolean",
            "title": "Is Outdated",
            "description": "Model không khớp analysis revisions hiện tại"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "dbml",
          "revision",
          "created_at",
          "updated_at",
          "is_outdated"
        ],
        "title": "DataModelResponse",
        "description": "Snapshot Data Model trả về cho Frontend editor."
      },
      "DataModelValidationIssueResponse": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ValidationIssueCode"
          },
          "table_name": {
            "type": "string",
            "title": "Table Name"
          },
          "column_name": {
            "type": "string",
            "title": "Column Name"
          },
          "severity": {
            "$ref": "#/components/schemas/ValidationSeverity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "code",
          "table_name",
          "column_name",
          "severity",
          "title",
          "description"
        ],
        "title": "DataModelValidationIssueResponse",
        "description": "Lỗi hoặc cảnh báo do ValidationEngine phát hiện."
      },
...
      "ErrorDetail": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Đường dẫn field không hợp lệ"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mô tả lỗi validation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Chi tiết validation của một trường request."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SandboxDbType": {
        "type": "string",
        "enum": [
          "POSTGRESQL",
          "BIGQUERY",
          "SNOWFLAKE",
          "MYSQL",
          "SQLITE",
          "SQLSERVER"
        ],
        "title": "SandboxDbType",
        "description": "Loại cơ sở dữ liệu Sandbox."
      },
...
      "UpdateDataModelRequest": {
        "properties": {
          "data_model_id": {
            "type": "string",
            "format": "uuid",
            "title": "Data Model Id",
            "description": "ID của Data Model cần cập nhật"
          },
          "dbml": {
            "type": "string",
            "maxLength": 1000000,
            "minLength": 1,
            "title": "Dbml",
            "description": "Toàn bộ snapshot DBML mới"
          },
          "base_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Base Revision",
            "description": "Revision mà client đã tải trước khi chỉnh sửa"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data_model_id",
          "dbml",
          "base_revision"
        ],
        "title": "UpdateDataModelRequest",
        "description": "Payload lưu trực tiếp DBML với revision gốc của client."
      },
...
      "ValidateDataModelRequest": {
        "properties": {
          "dbml": {
            "type": "string",
            "maxLength": 1000000,
            "minLength": 1,
            "title": "Dbml",
            "description": "DBML draft cần kiểm tra"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dbml"
        ],
        "title": "ValidateDataModelRequest",
        "description": "Payload kiểm tra DBML draft mà không ghi dữ liệu."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationIssueCode": {
        "type": "string",
        "enum": [
          "DBML_SYNTAX_INVALID",
          "TABLE_PRIMARY_KEY_MISSING",
          "TABLE_COLUMN_NAME_DUPLICATED",
          "RELATIONSHIP_DUPLICATED",
          "FACT_DIMENSION_RELATIONSHIP_MISSING"
        ],
        "title": "ValidationIssueCode",
        "description": "Mã ổn định định danh từng quy tắc validation Data Model."
      },
      "ValidationSeverity": {
        "type": "string",
        "enum": [
          "WARNING",
          "ERROR"
        ],
        "title": "ValidationSeverity",
        "description": "Mức độ của một vấn đề validation."
      }
    }
  }
}